stock predictor app/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pandas
numpy
scikit-learn
xgboost
yfinance
ta
python-dotenv
